fix: packeting dtos and moved part of price controller logic to service
