+++ src/main/java/rs/ac/uns/ftn/siit/team11/ProjectSIIT2023Team11/dto/AccommodationDTO/AccommodationAvailabilityDTO.java
@@ -1,5 +1,4 @@
-package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto;
-
+package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.AccommodationDTO;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,8 +7,9 @@ import java.time.LocalDate;
 
 @Data
 @AllArgsConstructor
-public class AvailabilityDTO {
-    private Long id;
+public class AccommodationAvailabilityDTO {
+
     private LocalDate startDate;
     private LocalDate endDate;
+
 }
