+++ src/main/java/rs/ac/uns/ftn/siit/team11/ProjectSIIT2023Team11/dto/AccommodationDTO/AccommodationCardDTO.java
@@ -1,4 +1,4 @@
-package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto;
+package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.AccommodationDTO;
 
 import jakarta.persistence.*;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.domain.Availability;
