+++ src/main/java/rs/ac/uns/ftn/siit/team11/ProjectSIIT2023Team11/dto/AccommodationDTO/AvailabilityDTO.java
@@ -1,4 +1,5 @@
-package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto;
+package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.AccommodationDTO;
+
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,9 +8,8 @@ import java.time.LocalDate;
 
 @Data
 @AllArgsConstructor
-public class AccommodationAvailabilityDTO {
-
+public class AvailabilityDTO {
+    private Long id;
     private LocalDate startDate;
     private LocalDate endDate;
-
 }
