+++ src/main/java/rs/ac/uns/ftn/siit/team11/ProjectSIIT2023Team11/dto/ReservationDTO/GuestReservationDTO.java
@@ -1,18 +1,17 @@
-package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto;
+package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.ReservationDTO;
 
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.util.ReservationStatus;
 
 import java.time.LocalDate;
-
 @Data
 @AllArgsConstructor
-public class ReservationForShowDTO {
-
+public class GuestReservationDTO {
     private Long id;
     private String accommodation;
-    private String guest;
     private LocalDate startDate;
     private LocalDate endDate;
     private int numberOfGuests;
