+++ src/main/java/rs/ac/uns/ftn/siit/team11/ProjectSIIT2023Team11/dto/ReservationDTO/OwnerReservationDTO.java
@@ -1,14 +1,14 @@
-package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto;
+package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.ReservationDTO;
 
-import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.util.ReservationStatus;
 
 import java.time.LocalDate;
+
 @Data
 @AllArgsConstructor
-public class ReservationDTO {
+public class OwnerReservationDTO {
     private Long id;
     private Long accommodation;
     private String guest;
@@ -17,4 +17,5 @@ public class ReservationDTO {
     private int numberOfGuests;
     private ReservationStatus status;
     private Double price;
+    private int cancelledReservations;
 }
