+++ src/main/java/rs/ac/uns/ftn/siit/team11/ProjectSIIT2023Team11/dto/ReservationDTO/ReservationForShowDTO.java
@@ -1,4 +1,4 @@
-package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto;
+package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.ReservationDTO;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,14 +8,14 @@ import java.time.LocalDate;
 
 @Data
 @AllArgsConstructor
-public class OwnerReservationDTO {
+public class ReservationForShowDTO {
+
     private Long id;
-    private Long accommodation;
+    private String accommodation;
     private String guest;
     private LocalDate startDate;
     private LocalDate endDate;
     private int numberOfGuests;
     private ReservationStatus status;
     private Double price;
-    private int cancelledReservations;
 }
