+++ src/main/java/rs/ac/uns/ftn/siit/team11/ProjectSIIT2023Team11/mapper/AvailabilityMapper.java
@@ -3,8 +3,8 @@ package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.mapper;
 import org.springframework.stereotype.Component;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.domain.Availability;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.domain.TimeSlot;
-import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.AccommodationAvailabilityDTO;
-import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.AvailabilityDTO;
+import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.AccommodationDTO.AccommodationAvailabilityDTO;
+import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.AccommodationDTO.AvailabilityDTO;
 
 @Component
 public class AvailabilityMapper {
