+++ src/main/java/rs/ac/uns/ftn/siit/team11/ProjectSIIT2023Team11/mapper/OwnerReservationMapper.java
@@ -2,8 +2,8 @@ package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.mapper;
 
 import org.springframework.stereotype.Component;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.domain.Reservation;
-import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.GuestReservationDTO;
-import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.OwnerReservationDTO;
+import GuestReservationDTO;
+import OwnerReservationDTO;
 
 @Component
 public class OwnerReservationMapper {
