+++ src/main/java/rs/ac/uns/ftn/siit/team11/ProjectSIIT2023Team11/service/AccommodationService.java
@@ -1,15 +1,14 @@
 package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Example;
 import org.springframework.stereotype.Service;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.domain.Accommodation;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.domain.Owner;
+import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.domain.Price;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.domain.User;
-import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.AccommodationDetailsDTO;
+import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.AccommodationDTO.AccommodationDetailsDTO;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.mapper.AccommodationMapper;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.repository.IAccommodationRepository;
-import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.repository.IUserRepository;
 
 import java.util.List;
 import java.util.Optional;
@@ -42,6 +41,17 @@ public class AccommodationService implements IAccommodationService{
         Owner owner = (Owner) user.get();
         Accommodation accommodation = AccommodationMapper.mapDetailsDtoToNewAccommodation(accommodationDetailsDTO,owner);
         return Optional.of(AccommodationMapper.mapToAccommodationDetailsDto(save(accommodation)));
+    }
+
+    @Override
+    public void deletePriceFromAllAccommodations(Price price) {
+        List<Accommodation> accommodations = findAll();
+        for(Accommodation accommodation : accommodations){
+            if(accommodation.containsPrice(price)){
+                accommodation.getPriceList().remove(price);
+                save(accommodation);
+            }
+        }
     }
 
 }
